backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
paramiko==3.3.1
cryptography==41.0.7
requests==2.31.0
ldap3==2.9.1
twilio==8.10.0
africastalking==1.2.8